+++ example/example.dart
@@ -5,7 +5,9 @@ import 'package:rpi_i2c/rpi_i2c.dart';
 import 'mpl3115a2.dart';
 
 main() async {
-  await readSensor(new Mpl3115a2(new RpiI2C()));
+  final i2c = new RpiI2C();
+  await readSensor(new Mpl3115a2(i2c));
+  i2c.dispose();
 }
 
 Future readSensor(Mpl3115a2 mpl3115a2) async {