+++ lib/i2c.dart
@@ -28,17 +28,28 @@ abstract class I2CDevice {
   /// Read a byte from register #[register].
   int readByte(int register);
 
+  /// Read [values].length # of bytes from the device (no register)
+  /// into [values] where the [values].length is between 1 and 32 inclusive.
+  /// Return the # of bytes read.
+  int readBytes(List<int> values);
+
   /// Write a byte to register #[register].
-  int writeByte(int register, int data);
+  void writeByte(int register, int value);
 }
 
 /// Exceptions thrown by I2C.
 class I2CException {
   final String message;
   final int address;
+  final int errorNumber;
 
-  I2CException(this.message, [this.address]);
+  I2CException(this.message, [this.address, this.errorNumber]);
 
   @override
-  String toString() => address != null ? '$message address: $address' : message;
+  String toString() {
+    String msg = message;
+    if (address != null) msg = '$msg, address: $address';
+    if (errorNumber != null) msg = '$msg, error: $errorNumber';
+    return 'I2CException: $msg';
+  }
 }