+++ tool/grind.dart
@@ -1,0 +1,8 @@
+import 'package:grinder/grinder.dart';
+
+main(List<String> args) => grind(args);
+
+@DefaultTask()
+void analyze() {
+  Analyzer.analyze(['.']);
+}