+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 0.1.1
+ * Add I2C readBytes support
+ * Add TTP229 touchpad example
+ * Add travic CI analysis
+ * Update readme and fix example code
+
 ## 0.1.0
  * Initial implementation